# build/swagger/swagger.yaml
basePath: /
definitions:
  api.HealthResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
  handlers.BatchInsertRequest:
    properties:
      vectors:
        items:
          $ref: '#/definitions/handlers.CreateVectorRequest'
        type: array
    required:
    - vectors
    type: object
  handlers.BatchInsertResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/handlers.BatchInsertResult'
        type: array
      inserted_count:
        type: integer
    type: object
  handlers.BatchInsertResult:
    properties:
      error:
        type: string
      id:
        type: string
    type: object
  handlers.CreateVectorRequest:
    properties:
      id:
        example: vec-001
        type: string
      metadata:
        type: object
      sparse_vector:
        type: object
      vector:
        items:
          type: number
        type: array
    required:
    - id
    - vector
    type: object
  handlers.DeleteResponse:
    properties:
      id:
        example: vec-001
        type: string
      status:
        example: deleted
        type: string
    type: object
  handlers.InsertResponse:
    properties:
      status:
        example: inserted
        type: string
    type: object
  handlers.SearchRequest:
    properties:
      filter:
        type: object
      k:
        example: 10
        type: integer
      sparse_vector:
        type: object
      vector:
        items:
          type: number
        type: array
    required:
    - vector
    type: object
  handlers.StatsResponse:
    properties:
      vector_count:
        example: 42
        type: integer
    type: object
  index.EngineInfo:
    properties:
      dimensions:
        example: 0
        type: integer
      distance_metric:
        example: cosine
        type: string
      index_type:
        example: brute
        type: string
      quantization:
        example: none
        type: string
      vector_count:
        example: 0
        type: integer
    type: object
  index.SearchResult:
    properties:
      id:
        example: vec-001
        type: string
      meta:
        type: object
      score:
        example: 0.97
        type: number
    type: object
  response.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
      success:
        example: false
        type: boolean
    type: object
  response.Response:
    properties:
      data: {}
      error:
        example: ""
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8000
info:
  contact: {}
  description: A vector database REST API for storing and querying vector embeddings.
  title: GoVec API
  version: "1.0"
paths:
  /api/v1/admin/flush:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Flush index to disk
      tags:
      - system
  /api/v1/info:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/index.EngineInfo'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get engine configuration info
      tags:
      - system
  /api/v1/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.StatsResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get index statistics
      tags:
      - system
  /api/v1/vectors:
    post:
      consumes:
      - application/json
      parameters:
      - description: Vector payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateVectorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.InsertResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Insert or update a vector
      tags:
      - vectors
  /api/v1/vectors/{id}:
    delete:
      parameters:
      - description: Vector ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.DeleteResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a vector by ID
      tags:
      - vectors
  /api/v1/vectors/batch:
    post:
      consumes:
      - application/json
      parameters:
      - description: Batch vector payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.BatchInsertRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.BatchInsertResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Batch insert vectors
      tags:
      - vectors
  /api/v1/vectors/search:
    post:
      consumes:
      - application/json
      parameters:
      - description: Search payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/index.SearchResult'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search for similar vectors
      tags:
      - vectors
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/api.HealthResponse'
              type: object
      summary: Health check
      tags:
      - system
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type 'Bearer ' followed by your API key.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"